feat(services): :scissors:  Improve `contentService` to flatten resolved links objects
Since link resolution object format does not work the same way in the REST API
as it does in contenful-js lib, a lot of wrangling might need to be required
to resolve links and abtrasct contenful data shapes from components.
Which means it's prob better to stick with contenful-js instead of vanila
REST calls.